backend/requirements.txt
aniso8601==9.0.1
APScheduler==3.10.4
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
cachelib==0.13.0
cffi==1.16.0
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==4.0.1
flask-restx==1.3.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
MarkupSafe==2.1.5
marshmallow==3.21.3
msgspec==0.18.6
packaging==24.1
pycparser==2.22
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-redis-lock==4.0.0
pytz==2024.1
redis==5.0.7
referencing==0.35.1
rpds-py==0.18.1
six==1.16.0
SQLAlchemy==2.0.30
typing_extensions==4.12.1
tzdata==2024.1
tzlocal==5.2
Werkzeug==3.0.3
